refactor: change date picker components and example stories to separated stories folder
