+++ .storybook/preview.tsx
@@ -1,7 +1,6 @@
 import '../src/styles/main.css'
 
 import type { Preview } from '@storybook/react'
-import { DatePickerProvider } from '../src/components/DatePickerProvider'
 
 const preview: Preview = {
   parameters: {
