+++ package.json
@@ -1,7 +1,7 @@
 {
   "name": "react-modular-datepicker",
   "private": true,
-  "version": "0.0.0",
+  "version": "0.1.0",
   "type": "module",
   "scripts": {
     "build": "tsc -b && vite build",
