+++ stories/DatePickerProvider.stories.tsx
@@ -1,9 +1,9 @@
 ﻿import { Meta, StoryObj } from '@storybook/react'
-import { DatePickerProvider } from '.'
-import { DatePickerProviderProps } from '../../types'
-import { Calendar } from '../Calendar'
-import { Header, Button, MonthLabel } from '../Header'
-import { DateSelectExample } from '../../../.storybook/components/DateSelectExample'
+import { DatePickerProvider } from '../src/components/DatePickerProvider'
+import { DatePickerProviderProps } from '../src/types'
+import { Calendar } from '../src/components/Calendar'
+import { Header, Button, MonthLabel } from '../src/components/Header'
+import { DateSelectExample } from '../.storybook/components/DateSelectExample'
 import { ChevronLeft, ChevronRight, Calendar1, CalendarDays, CalendarRange } from 'lucide-react'
 import dayjs from 'dayjs'
 import 'dayjs/locale/pt-br'
